+++ kani-compiler/src/codegen_cprover_gotoc/codegen/mod.rs
@@ -17,6 +17,7 @@ mod statement;
 mod static_var;
 
 // Visible for all codegen module.
+pub mod contract;
 mod ty_stable;
 pub(super) mod typ;
